requirements.txt
streamlit>=1.28.0
pymupdf>=1.23.0
google-generativeai
Pillow>=10.0.0
regex>=2023.0.0
